% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanse_html_tags.R
\name{clean_html_tags}
\alias{clean_html_tags}
\title{Clean out HTML tags}
\usage{
clean_html_tags(stringVec)
}
\arguments{
\item{stringVec}{}
}
\value{
clean string or list of strings
}
\description{
Clean out HTML tags
}
