% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conntect_to_database.R
\name{connectToDatabase}
\alias{connectToDatabase}
\title{Connect to Database}
\usage{
connectToDatabase(datacenter)
}
\arguments{
\item{datacenter}{c('dev','QA','Virginia','Frankfurt','USbackup')}
}
\value{
TRUE if datacenter is on the list FALSE if not.
}
\description{
Connects to an X-Ray database (Postgres)
}
