% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data_frame_percentage.R
\name{format_data_frame_percentage}
\alias{format_data_frame_percentage}
\title{Format data.frame with Percentages}
\usage{
format_data_frame_percentage(df, na.string = "", digits = 2)
}
\arguments{
\item{df}{the data.frame to format}

\item{na.string}{what to replace NA's with}

\item{digits}{number of digits to round to}
}
\value{
a formatted data.frame
}
\description{
Will format all numbers in a data.frame as percentage. 
And replace NA with string specified in *na.string* parameter.
}
