% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_percent.R
\name{format_percent}
\alias{format_percent}
\title{Format data.frame with Percentage}
\usage{
format_percent(df, na.string = "")
}
\arguments{
\item{df}{}

\item{na.string}{what to replace NA's with}
}
\value{
a formatted data.frame
}
\description{
Will format all numbers in a data.frame as percentage. And replace NA with string specified in *replace* parameter.
}
