% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comma_separated_list_with_and.R
\name{get_comma_separated_list_with_and}
\alias{get_comma_separated_list_with_and}
\title{Format Vector as Comma Separated}
\usage{
get_comma_separated_list_with_and(vec, and = "and",
  harvard.comma = FALSE)
}
\arguments{
\item{vec}{A character vector}

\item{and}{the 'and' word. i.e in spansh 'y'}

\item{harvard.comma}{should a harvard comma be inserted}
}
\value{
a formatted character string.
}
\description{
Format Vector as Comma Separated
}
