% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_with_density_line.R
\name{getHistogramWithDensityLine}
\alias{getHistogramWithDensityLine}
\title{Generate Histogram with Density Line}
\usage{
getHistogramWithDensityLine(myData, binwidth = 0.1, variable = "x")
}
\arguments{
\item{myData}{data.frame with raw data}

\item{binwidth}{binwidth for the histogram}

\item{variable}{variable to be plotted}
}
\value{
a ggplot object histogram whith densityline overplotted
}
\description{
Generate Histogram with Density Line
}
