% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_lm.R
\name{latex.lm}
\alias{latex.lm}
\title{Print linear model object as latex formula}
\usage{
latex.lm(object, file = "", math.env = c("$", "$"), estimates = "none",
  abbreviate = TRUE, abbrev.length = 8, digits = 3)
}
\arguments{
\item{object}{the linear object}

\item{file}{file to write to. Leave blank to print to screen (Default)}

\item{math.env}{Surrounding tags $ and $ by default}

\item{estimates}{c("none","inline","separate") for parameter estimates}

\item{abbreviate}{Should terms be abbreviated}

\item{abbrev.length}{Length of the abbrevitation}

\item{digits}{Precision}
}
\value{
A latex rendition of the model
}
\description{
Print linear model object as latex formula
}
