% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple Graphics in One plot}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{The plots you want to include}

\item{plotlist}{The plots you want to include}

\item{file}{Filename}

\item{cols}{number of columns}

\item{layout}{}
}
\description{
Place an arbitrary number of plot in rows and columns within the space of the same plot
}
