% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num2words.R
\name{num2words}
\alias{num2words}
\title{Convert Number to Word}
\usage{
num2words(x, first.cap = FALSE)
}
\arguments{
\item{x}{}

\item{first.cap}{Should the first word be capitalized (i.e. start of a sentence)}
}
\value{
a string
}
\description{
Convert a number to word.
}
