% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_summary.R
\name{prettySummary}
\alias{prettySummary}
\title{Pretty Summary}
\usage{
prettySummary(input, digits = 3)
}
\arguments{
\item{input}{The Object to Pretty Summarize}

\item{digits}{Significant digits}
}
\description{
Pretty Summary for various objects.
}
