% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_seconds_as_duration.R
\name{seconds_as_duration}
\alias{seconds_as_duration}
\title{Format Seconds Nicely}
\usage{
seconds_as_duration(x, format = "\%02d:\%02d:\%02d")
}
\arguments{
\item{x}{Vector of seconds (durations). Numeric or coercible to numeric}

\item{format}{for consumption by sprintf}
}
\value{
a vector of characters
}
\description{
Format Seconds Nicely
}
\examples{
seconds_as_duration(c(100,200,3024, 16254))
seconds_as_duration(c(100,200,3024, 16254), format=NULL)
seconds_as_duration(c(100,200,3024, 16254), format='\%02dh\%02dm\%02ds')
}
