% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states_by_population}
\alias{states_by_population}
\title{State by Population}
\format{A data frame with 52 rows and 3 variables:
\describe{
  \item{State}{State Name}
  \item{abb}{Abbreviated (2 letter) designation}
  \item{Population}{The population in 1000s}
}}
\usage{
states_by_population
}
\description{
A dataset containing population (in thousands) for each US state plus PR and DC.
}
\keyword{datasets}
